requirements.txt
apache-airflow==2.7.0
pandas==1.5.3
requests==2.28.2
pyarrow==18.1.0
fastparquet==0.8.3
pytest==7.4.2
pyspark==3.4.0